improved deploy config. now use a single configuration file

+++ CompileAndDeploy.php
@@ -1,37 +1,46 @@
 #!/usr/bin/php
 <?php
 $debug = false;
+$cleanedArgs = array();
+foreach ($argv as $arg) {
+        if ($arg == "debug") {
+                $debug = true;
+        } else {
+                $cleanedArgs[] = $arg;
+        }
+}
+error_reporting(E_ERROR);
 include_once __DIR__.'/Helper.php';
 $config = Helper::getConfig();
-if (in_array("debug", $argv)) $debug = true;
 
-
-if (count($argv) > 1) {
-	if (in_array($argv[1], explode(",", $config['deployments']))) {
-		$deployConfig = $config[$argv[1]];
-		if (is_array($deployConfig) && count($deployConfig)) {
-			compileAndDeploy($deployConfig);
-		}
-	}
+if (count($cleanedArgs) > 1) {
+        if (is_array($config['deployTargets'][$cleanedArgs[1]])) {
+                foreach ($config['deployTargets'][$cleanedArgs[1]] as $deployConfig) {
+                	echo "\n== deploying to target ".$deployConfig['target']."\n";
+                    compileAndDeploy($deployConfig);
+                }
+        }
+} else {
+        echo "not enough arguments: \n ./CompileAndDeploy <branch>\n";
 }
-
-
 
 
 function compileAndDeploy($deployConfig) {
 	global $argv;
-	chdir($deployConfig['git']);
+	chdir($deployConfig['src']);
 	echo "pulling\n";
  	pull();
  	echo "packaging\n";
  	package();
 	$warFile = getWarFile();
 	$yasguiDir = unzipWarFile($warFile);
-	echo "updating config";
+	echo "updating config\n";
 	updateConfig($yasguiDir, $deployConfig);
-	echo "deploying to tomcat";
+	echo "deploying to tomcat\n";
 	deployToTomcat($yasguiDir, $deployConfig);
- 	Helper::sendMail("Succesfully deployed YASGUI as ".$argv[1], "Succesfully deployed YASGUI as ".$argv[1]);
+	$subject = "Succesfully deployed YASGUI as ".$argv[1]." on ".gethostname();
+	$body = $subject."\nmode: ".$deployConfig['target'];
+ 	Helper::sendMail($succesText, $succesText);
 }
 
 function pull() {
@@ -44,11 +53,11 @@ function pull() {
 
 function package() {
 	global $argv;
-	$succes = shell_exec("mvn clean 2> errorOutput.txt");
-	if (!$succes) {
-		Helper::mailError(__FILE__, __LINE__, "Unable to compile ".$argv[1]." project: \n".file_get_contents("errorOutput.txt"));
-	}
-	if ($succes) $succes = shell_exec("mvn package 2> errorOutput.txt");
+// 	$succes = shell_exec("mvn clean 2> errorOutput.txt");
+// 	if (!$succes) {
+// 		Helper::mailError(__FILE__, __LINE__, "Unable to compile ".$argv[1]." project: \n".file_get_contents("errorOutput.txt"));
+// 	}
+	$succes = shell_exec("mvn package 2> errorOutput.txt");
 	if (!$succes || strpos($succes, "BUILD FAILURE")) {
 		Helper::mailError(__FILE__, __LINE__, "Unable to compile ".$argv[1]." project: \n".file_get_contents("errorOutput.txt")."\n".$succes);
 	}
@@ -90,18 +99,14 @@ function getUpdatedConfig($dir, $deployConfig) {
 	if (!file_exists($jsonConfig)) {
 		Helper::mailError(__FILE__, __LINE__, "No config file in unzipped war file (".$jsonConfig.")");
 	}
-	$overWriteJsonConfig = $deployConfig['yasguiConfig'];
-	if (!file_exists($overWriteJsonConfig)) {
-		Helper::mailError(__FILE__, __LINE__, "No json config file to apply to yasgui (".$overWriteJsonConfig.")");
-	}
-	
 	$jsonConfigArray = json_decode(file_get_contents($jsonConfig), true);
 	if (!(is_array($jsonConfigArray) && count($jsonConfigArray))) {
 		Helper::mailError(__FILE__, __LINE__, "Unable to parse file as json (".$jsonConfig.")");
 	}
-	$overWriteJsonConfigArray = json_decode(file_get_Contents($overWriteJsonConfig), true);
+	
+	$overWriteJsonConfigArray = $deployConfig['overwriteConfig'];
 	if (!(is_array($overWriteJsonConfigArray) && count($jsonConfigArray))) {
-		Helper::mailError(__FILE__, __LINE__, "Unable to parse file as json (".$overWriteJsonConfig.")");
+		Helper::mailError(__FILE__, __LINE__, "No json config file to change (".$overWriteJsonConfig.")");
 	}
 	return array_replace_recursive($jsonConfigArray, $overWriteJsonConfigArray);
 }
@@ -114,8 +119,7 @@ function generateRandomString($length = 10) {
 	return $randomString;
 }
 function deployToTomcat($yasguiDir, $deployConfig) {
-	global $config;
-	$to = $deployConfig['tomcat'];
+	$to = $deployConfig['target'];
 	$tmpDir = sys_get_temp_dir()."/".generateRandomString();
 	shell_exec("mkdir ".$tmpDir);
 	/**
@@ -123,8 +127,10 @@ function deployToTomcat($yasguiDir, $deployConfig) {
 	 */
 	//be very sure we arent deleting other stuff:
 	if (strlen($to) && file_exists($to) && strpos($to, "tomcat")) {
-		//first copy cache dir (we want to save that one)
-		shell_exec("mv ".$to."/cache ".$tmpDir);
+		if (file_exists($to."/cache")) {
+			//first copy cache dir (we want to save that one)
+			shell_exec("mv ".$to."/cache ".$tmpDir);
+		}
 		
 		shell_exec("rm -rf ".$to);
 		//all files created by YASGUI won't be deleted... (owned by tomcat, not apache)
@@ -155,7 +161,9 @@ function deployToTomcat($yasguiDir, $deployConfig) {
 	 * restore cach dir
 	 */
 	shell_exec("rm -r ".$to."/cache");
-	shell_exec("mv ".$tmpDir."/cache ".$to);
+	if (file_exists($tmpDir."/cache")) {
+		shell_exec("mv ".$tmpDir."/cache ".$to);
+	}
 	
 }
